Use lua_rawsetfield; add Luau defs
Replace lua_setfield with lua_rawsetfield in reflection code to write fields directly (avoid metamethods). ) Also update EditorSettings recent projects, add test sources (ECSTests.cpp, LuaTests.cpp), and touch various engine headers/sources to support these changes.

+++ Engine/Source/Runtime/Scripting/Lua/Traits.cpp
@@ -6,8 +6,8 @@ namespace Lumina::Lua
 {
     uint16 FTypeIndex::Next()
     {
-        static uint16_t GNextID = 1;
+        static uint16 GNextID = 1;
         DEBUG_ASSERT(GNextID <= eastl::numeric_limits<uint16>::max(), "Hit the maximum number of allowed userdata!");
-        return static_cast<uint16>(GNextID++);
+        return GNextID++;
     }
 }
